+++ bin/tab_completion.sh
@@ -13,7 +13,7 @@ __filedirs_all()
 }
 
 _mde_echo_version() {
-  echo "2.4.1"
+  echo "2.5.0"
 }
 
 _mde() {